Move validate functions
